+++ crates/language_model_core/Cargo.toml
@@ -10,12 +10,12 @@ workspace = true
 
 [lib]
 path = "src/language_model_core.rs"
-doctest = false
 
 [dependencies]
 anyhow.workspace = true
 async-lock.workspace = true
 cloud_llm_client.workspace = true
+collections.workspace = true
 futures.workspace = true
 gpui_shared_string.workspace = true
 http_client.workspace = true
