+++ crates/language_models/src/provider/open_ai_compatible.rs
@@ -3,6 +3,7 @@ use credentials_provider::CredentialsProvider;
 use futures::{FutureExt, StreamExt, future::BoxFuture};
 use gpui::{App, AppContext, AsyncApp, Entity, Task};
 use http_client::{CustomHeaders, HttpClient};
+use language_model::chat_completion::ChatCompletionEventMapper;
 use language_model::{
     AuthenticateError, IconOrSvg, LanguageModel, LanguageModelCompletionError,
     LanguageModelCompletionEvent, LanguageModelEffortLevel, LanguageModelId, LanguageModelName,
@@ -23,9 +24,7 @@ use crate::provider::api_compatible::{
     ApiCompatibleProviderConfigurationView, ApiCompatibleProviderSettings,
     ApiCompatibleProviderState,
 };
-use crate::provider::open_ai::{
-    OpenAiEventMapper, OpenAiResponseEventMapper, into_open_ai, into_open_ai_response,
-};
+use crate::provider::open_ai::{OpenAiResponseEventMapper, into_open_ai, into_open_ai_response};
 pub use settings::OpenAiCompatibleAvailableModel as AvailableModel;
 pub use settings::OpenAiCompatibleModelCapabilities as ModelCapabilities;
 
@@ -436,7 +435,7 @@ impl LanguageModel for OpenAiCompatibleLanguageModel {
             let completions = self.stream_completion(request, cx);
             let executor = cx.background_executor().clone();
             async move {
-                let mapper = OpenAiEventMapper::new();
+                let mapper = ChatCompletionEventMapper::new();
                 Ok(language_model::stream_in_background(
                     mapper.map_stream(completions.await?).boxed(),
                     executor,
