+++ crates/language_models/src/provider/vercel_ai_gateway.rs
@@ -463,7 +463,7 @@ impl LanguageModel for VercelAiGatewayLanguageModel {
         let completions = self.stream_open_ai(request, cx);
         let executor = cx.background_executor().clone();
         async move {
-            let mapper = crate::provider::open_ai::OpenAiEventMapper::new();
+            let mapper = language_model::chat_completion::ChatCompletionEventMapper::new();
             Ok(language_model::stream_in_background(
                 mapper.map_stream(completions.await?).boxed(),
                 executor,
