+++ crates/open_ai/src/completion.rs
@@ -25,8 +25,8 @@ use crate::responses::{
     provider_compaction_items, provider_compaction_state_from_items,
 };
 use crate::{
-    FunctionContent, FunctionDefinition, ImageUrl, MessagePart, ReasoningEffort,
-    ResponseStreamEvent, ServiceTier, ToolCall, ToolCallContent,
+    FunctionContent, FunctionDefinition, ImageUrl, MessagePart, ReasoningEffort, ServiceTier,
+    ToolCall, ToolCallContent,
 };
 
 const RESPONSE_MESSAGE_PHASE_COMMENTARY: &str = "commentary";
@@ -727,268 +727,6 @@ fn add_message_content_part(
             });
         }
     }
-}
-
-/// Accumulates structured reasoning metadata from compatible providers.
-///
-/// Array entries are matched by `index` and then `id`. Fragmented `text`,
-/// `summary`, and `data` fields are concatenated while other non-null fields
-/// replace their previous values.
-///
-/// # Examples
-///
-/// ```
-/// use open_ai::completion::ReasoningDetailsAccumulator;
-/// use serde_json::json;
-///
-/// let mut accumulator = ReasoningDetailsAccumulator::default();
-/// accumulator.push(json!([{"index": 0, "text": "first "}]));
-/// let details = accumulator
-///     .push(json!([{"index": 0, "text": "second"}]))
-///     .expect("non-empty reasoning details");
-///
-/// assert_eq!(details[0]["text"], "first second");
-/// ```
-#[derive(Debug, Default)]
-pub struct ReasoningDetailsAccumulator {
-    accumulated: Option<serde_json::Value>,
-}
-
-impl ReasoningDetailsAccumulator {
-    /// Merges `chunk` and returns the updated metadata snapshot.
-    ///
-    /// `null` and empty arrays do not replace previously accumulated metadata
-    /// and return `None`.
-    pub fn push(&mut self, chunk: serde_json::Value) -> Option<serde_json::Value> {
-        match chunk {
-            serde_json::Value::Null => None,
-            serde_json::Value::Array(chunks) if chunks.is_empty() => None,
-            serde_json::Value::Array(chunks) => {
-                let mut details = match self.accumulated.take() {
-                    Some(serde_json::Value::Array(details)) => details,
-                    _ => Vec::new(),
-                };
-                for chunk in chunks {
-                    merge_reasoning_detail(&mut details, chunk);
-                }
-                let accumulated = serde_json::Value::Array(details);
-                self.accumulated = Some(accumulated.clone());
-                Some(accumulated)
-            }
-            chunk => {
-                self.accumulated = Some(chunk.clone());
-                Some(chunk)
-            }
-        }
-    }
-}
-
-pub struct OpenAiEventMapper {
-    tool_calls_by_index: HashMap<usize, RawToolCall>,
-    reasoning_details: ReasoningDetailsAccumulator,
-}
-
-impl OpenAiEventMapper {
-    pub fn new() -> Self {
-        Self {
-            tool_calls_by_index: HashMap::default(),
-            reasoning_details: ReasoningDetailsAccumulator::default(),
-        }
-    }
-
-    pub fn map_stream<E>(
-        mut self,
-        events: Pin<Box<dyn Send + Stream<Item = Result<ResponseStreamEvent, E>>>>,
-    ) -> impl Stream<Item = Result<LanguageModelCompletionEvent, LanguageModelCompletionError>>
-    where
-        E: Into<LanguageModelCompletionError>,
-    {
-        events.flat_map(move |event| {
-            futures::stream::iter(match event {
-                Ok(event) => self.map_event(event),
-                Err(error) => vec![Err(error.into())],
-            })
-        })
-    }
-
-    pub fn map_event(
-        &mut self,
-        event: ResponseStreamEvent,
-    ) -> Vec<Result<LanguageModelCompletionEvent, LanguageModelCompletionError>> {
-        let mut events = Vec::new();
-        if let Some(token_usage) = event.usage.as_ref().and_then(|usage| usage.token_usage()) {
-            events.push(Ok(LanguageModelCompletionEvent::UsageUpdate(token_usage)));
-        }
-
-        let Some(choice) = event.choices.first() else {
-            return events;
-        };
-
-        if let Some(delta) = choice.delta.as_ref() {
-            if let Some(reasoning_details) = delta.reasoning_details.clone()
-                && let Some(reasoning_details) = self.reasoning_details.push(reasoning_details)
-            {
-                events.push(Ok(LanguageModelCompletionEvent::ReasoningDetails(
-                    reasoning_details,
-                )));
-            }
-            if let Some(reasoning) = delta.reasoning.clone() {
-                push_thinking_event(reasoning, &mut events);
-            }
-            if let Some(reasoning_content) = delta.reasoning_content.clone() {
-                push_thinking_event(reasoning_content, &mut events);
-            }
-            if let Some(content) = delta.content.clone() {
-                if !content.is_empty() {
-                    events.push(Ok(LanguageModelCompletionEvent::Text(content)));
-                }
-            }
-
-            if let Some(tool_calls) = delta.tool_calls.as_ref() {
-                for tool_call in tool_calls {
-                    let entry = self.tool_calls_by_index.entry(tool_call.index).or_default();
-
-                    if let Some(tool_id) = tool_call.id.clone()
-                        && !tool_id.is_empty()
-                    {
-                        entry.id = tool_id;
-                    }
-
-                    if let Some(function) = tool_call.function.as_ref() {
-                        if let Some(name) = function.name.clone()
-                            && !name.is_empty()
-                        {
-                            entry.name = name;
-                        }
-
-                        if let Some(arguments) = function.arguments.clone() {
-                            entry.arguments.push_str(&arguments);
-                        }
-
-                        if let Some(thought_signature) = function.thought_signature.clone() {
-                            entry.thought_signature = Some(thought_signature);
-                        }
-                    }
-
-                    if !entry.id.is_empty() && !entry.name.is_empty() {
-                        if let Ok(input) = serde_json::from_str::<serde_json::Value>(
-                            &fix_streamed_json(&entry.arguments),
-                        ) {
-                            events.push(Ok(LanguageModelCompletionEvent::ToolUse(
-                                LanguageModelToolUse {
-                                    id: entry.id.clone().into(),
-                                    name: entry.name.as_str().into(),
-                                    is_input_complete: false,
-                                    input: LanguageModelToolUseInput::Json(input),
-                                    raw_input: entry.arguments.clone(),
-                                    thought_signature: entry.thought_signature.clone(),
-                                },
-                            )));
-                        }
-                    }
-                }
-            }
-        }
-
-        match choice.finish_reason.as_deref() {
-            Some("stop") => {
-                events.push(Ok(LanguageModelCompletionEvent::Stop(StopReason::EndTurn)));
-            }
-            Some("tool_calls") => {
-                events.extend(self.tool_calls_by_index.drain().map(|(_, tool_call)| {
-                    match parse_tool_arguments(&tool_call.arguments) {
-                        Ok(input) => Ok(LanguageModelCompletionEvent::ToolUse(
-                            LanguageModelToolUse {
-                                id: tool_call.id.clone().into(),
-                                name: tool_call.name.as_str().into(),
-                                is_input_complete: true,
-                                input: LanguageModelToolUseInput::Json(input),
-                                raw_input: tool_call.arguments.clone(),
-                                thought_signature: tool_call.thought_signature.clone(),
-                            },
-                        )),
-                        Err(error) => Ok(LanguageModelCompletionEvent::ToolUseJsonParseError {
-                            id: tool_call.id.into(),
-                            tool_name: tool_call.name.into(),
-                            raw_input: tool_call.arguments.clone().into(),
-                            json_parse_error: error.to_string(),
-                        }),
-                    }
-                }));
-
-                events.push(Ok(LanguageModelCompletionEvent::Stop(StopReason::ToolUse)));
-            }
-            Some("length") => {
-                events.push(Ok(LanguageModelCompletionEvent::Stop(
-                    StopReason::MaxTokens,
-                )));
-            }
-            Some(stop_reason) => {
-                log::error!("Unexpected chat completion stop_reason: {stop_reason:?}",);
-                events.push(Ok(LanguageModelCompletionEvent::Stop(StopReason::EndTurn)));
-            }
-            None => {}
-        }
-
-        events
-    }
-}
-
-fn push_thinking_event(
-    text: String,
-    events: &mut Vec<Result<LanguageModelCompletionEvent, LanguageModelCompletionError>>,
-) {
-    if !text.is_empty() {
-        events.push(Ok(LanguageModelCompletionEvent::Thinking {
-            text,
-            signature: None,
-        }));
-    }
-}
-
-fn merge_reasoning_detail(details: &mut Vec<serde_json::Value>, chunk: serde_json::Value) {
-    let index = chunk.get("index").and_then(serde_json::Value::as_u64);
-    let target_index = index
-        .and_then(|index| {
-            details.iter().position(|detail| {
-                detail.get("index").and_then(serde_json::Value::as_u64) == Some(index)
-            })
-        })
-        .or_else(|| {
-            let id = chunk.get("id").and_then(serde_json::Value::as_str)?;
-            details
-                .iter()
-                .position(|detail| detail.get("id").and_then(serde_json::Value::as_str) == Some(id))
-        });
-    let Some(target_index) = target_index else {
-        details.push(chunk);
-        return;
-    };
-    let (Some(target), Some(chunk)) = (details[target_index].as_object_mut(), chunk.as_object())
-    else {
-        return;
-    };
-    for (key, value) in chunk {
-        if matches!(key.as_str(), "text" | "summary" | "data")
-            && let Some(fragment) = value.as_str()
-            && let Some(existing) = target.get(key).and_then(serde_json::Value::as_str)
-        {
-            target.insert(
-                key.clone(),
-                serde_json::Value::String(format!("{existing}{fragment}")),
-            );
-        } else if !value.is_null() {
-            target.insert(key.clone(), value.clone());
-        }
-    }
-}
-
-#[derive(Default)]
-struct RawToolCall {
-    id: String,
-    name: String,
-    arguments: String,
-    thought_signature: Option<String>,
 }
 
 pub struct OpenAiResponseEventMapper {
@@ -1761,9 +1499,6 @@ mod tests {
     use serde_json::json;
 
     use super::*;
-    use crate::{
-        ChoiceDelta, FunctionChunk, ResponseMessageDelta, ResponseStreamEvent, ToolCallChunk, Usage,
-    };
 
     fn map_response_events(events: Vec<ResponsesStreamEvent>) -> Vec<LanguageModelCompletionEvent> {
         block_on(async {
@@ -1775,17 +1510,6 @@ mod tests {
                 .map(Result::unwrap)
                 .collect()
         })
-    }
-
-    fn map_completion_events(
-        events: Vec<ResponseStreamEvent>,
-    ) -> Vec<LanguageModelCompletionEvent> {
-        let mut mapper = OpenAiEventMapper::new();
-        let mut all_events = Vec::new();
-        for event in events {
-            all_events.extend(mapper.map_event(event));
-        }
-        all_events.into_iter().filter_map(|e| e.ok()).collect()
     }
 
     fn response_item_message(id: &str) -> ResponseOutputItem {
@@ -4308,408 +4032,6 @@ mod tests {
                 {"role": "tool", "content": "result", "tool_call_id": "call-1"}
             ])
         );
-    }
-
-    #[test]
-    fn stream_maps_reasoning() {
-        let events = map_completion_events(vec![ResponseStreamEvent {
-            choices: vec![ChoiceDelta {
-                index: 0,
-                delta: Some(ResponseMessageDelta {
-                    content: None,
-                    reasoning: Some("thinking".into()),
-                    tool_calls: None,
-                    reasoning_content: None,
-                    reasoning_details: None,
-                }),
-                finish_reason: None,
-            }],
-            usage: None,
-        }]);
-
-        assert_eq!(
-            events,
-            vec![LanguageModelCompletionEvent::Thinking {
-                text: "thinking".into(),
-                signature: None,
-            }]
-        );
-    }
-
-    #[test]
-    fn stream_maps_length_finish_reason_to_max_tokens_stop() {
-        let events = map_completion_events(vec![ResponseStreamEvent {
-            choices: vec![ChoiceDelta {
-                index: 0,
-                delta: None,
-                finish_reason: Some("length".into()),
-            }],
-            usage: None,
-        }]);
-
-        assert_eq!(
-            events,
-            vec![LanguageModelCompletionEvent::Stop(StopReason::MaxTokens)]
-        );
-    }
-
-    #[test]
-    fn chunk_without_choices_or_usage_maps_to_no_events() {
-        let mut mapper = OpenAiEventMapper::new();
-        assert!(mapper.map_event(ResponseStreamEvent::default()).is_empty());
-    }
-
-    #[test]
-    fn usage_update_precedes_text_and_stop_events_from_the_same_chunk() {
-        let mut mapper = OpenAiEventMapper::new();
-        let events = mapper.map_event(ResponseStreamEvent {
-            choices: vec![ChoiceDelta {
-                index: 0,
-                delta: Some(ResponseMessageDelta {
-                    content: Some("Hello!".to_string()),
-                    ..Default::default()
-                }),
-                finish_reason: Some("stop".to_string()),
-            }],
-            usage: Some(Usage {
-                prompt_tokens: Some(11),
-                completion_tokens: Some(7),
-                total_tokens: Some(18),
-                prompt_tokens_details: None,
-            }),
-        });
-
-        assert!(matches!(
-            events.as_slice(),
-            [
-                Ok(LanguageModelCompletionEvent::UsageUpdate(TokenUsage {
-                    input_tokens: 11,
-                    output_tokens: 7,
-                    cache_creation_input_tokens: 0,
-                    cache_read_input_tokens: 0,
-                })),
-                Ok(LanguageModelCompletionEvent::Text(text)),
-                Ok(LanguageModelCompletionEvent::Stop(StopReason::EndTurn)),
-            ] if text == "Hello!"
-        ));
-    }
-
-    #[test]
-    fn usage_update_precedes_tool_use_and_stop_events_from_the_same_chunk() {
-        let mut mapper = OpenAiEventMapper::new();
-        let events = mapper.map_event(ResponseStreamEvent {
-            choices: vec![ChoiceDelta {
-                index: 0,
-                delta: Some(ResponseMessageDelta {
-                    tool_calls: Some(vec![ToolCallChunk {
-                        index: 0,
-                        id: Some("tool-call-id".to_string()),
-                        function: Some(FunctionChunk {
-                            name: Some("test_tool".to_string()),
-                            arguments: Some(r#"{"value":1}"#.to_string()),
-                            thought_signature: None,
-                        }),
-                    }]),
-                    ..Default::default()
-                }),
-                finish_reason: Some("tool_calls".to_string()),
-            }],
-            usage: Some(Usage {
-                prompt_tokens: Some(13),
-                completion_tokens: Some(5),
-                total_tokens: Some(18),
-                prompt_tokens_details: None,
-            }),
-        });
-
-        assert!(matches!(
-            events.as_slice(),
-            [
-                Ok(LanguageModelCompletionEvent::UsageUpdate(TokenUsage {
-                    input_tokens: 13,
-                    output_tokens: 5,
-                    cache_creation_input_tokens: 0,
-                    cache_read_input_tokens: 0,
-                })),
-                Ok(LanguageModelCompletionEvent::ToolUse(LanguageModelToolUse {
-                    is_input_complete: false,
-                    ..
-                })),
-                Ok(LanguageModelCompletionEvent::ToolUse(LanguageModelToolUse {
-                    id,
-                    name,
-                    is_input_complete: true,
-                    ..
-                })),
-                Ok(LanguageModelCompletionEvent::Stop(StopReason::ToolUse)),
-            ] if id.to_string() == "tool-call-id" && name.as_ref() == "test_tool"
-        ));
-    }
-
-    #[test]
-    fn stream_merges_reasoning_details_and_maps_compatible_usage_and_signatures() {
-        let response_events = serde_json::from_value(json!([
-            {
-                "choices": [{
-                    "index": 0,
-                    "delta": {
-                        "reasoning_details": [{
-                            "id": "reasoning-1",
-                            "index": 0,
-                            "type": "reasoning.text",
-                            "text": "first "
-                        }],
-                        "tool_calls": [{
-                            "index": 0,
-                            "id": "call-1",
-                            "function": {
-                                "name": "search",
-                                "arguments": "{",
-                                "thought_signature": "signature"
-                            }
-                        }]
-                    },
-                    "finish_reason": null
-                }],
-                "usage": null
-            },
-            {
-                "choices": [{
-                    "index": 0,
-                    "delta": {
-                        "reasoning_details": [{
-                            "id": "reasoning-1",
-                            "index": 0,
-                            "text": "second"
-                        }],
-                        "tool_calls": [{
-                            "index": 0,
-                            "function": {
-                                "arguments": "}"
-                            }
-                        }]
-                    },
-                    "finish_reason": "tool_calls"
-                }],
-                "usage": null
-            },
-            {
-                "choices": [],
-                "usage": {
-                    "prompt_tokens": 10000,
-                    "completion_tokens": 500,
-                    "total_tokens": 10500,
-                    "prompt_tokens_details": {
-                        "cached_tokens": 6000,
-                        "cache_write_tokens": 1000
-                    }
-                }
-            }
-        ]))
-        .expect("valid compatible Chat Completions events");
-        let events = map_completion_events(response_events);
-
-        assert!(events.iter().any(|event| {
-            matches!(
-                event,
-                LanguageModelCompletionEvent::ReasoningDetails(details)
-                    if details[0]["text"] == "first second"
-            )
-        }));
-        assert!(events.iter().any(|event| {
-            matches!(
-                event,
-                LanguageModelCompletionEvent::ToolUse(tool_use)
-                    if tool_use.is_input_complete
-                        && tool_use.thought_signature.as_deref() == Some("signature")
-            )
-        }));
-        assert!(events.iter().any(|event| {
-            matches!(
-                event,
-                LanguageModelCompletionEvent::UsageUpdate(TokenUsage {
-                    input_tokens: 3_000,
-                    output_tokens: 500,
-                    cache_creation_input_tokens: 1_000,
-                    cache_read_input_tokens: 6_000,
-                })
-            )
-        }));
-    }
-
-    #[test]
-    fn reasoning_details_accumulator_replaces_an_incompatible_previous_shape() {
-        let mut accumulator = ReasoningDetailsAccumulator::default();
-        assert_eq!(
-            accumulator.push(json!({"summary": "provider-defined"})),
-            Some(json!({"summary": "provider-defined"}))
-        );
-        assert_eq!(
-            accumulator.push(json!([{"index": 0, "text": "reasoning"}])),
-            Some(json!([{"index": 0, "text": "reasoning"}]))
-        );
-    }
-
-    // OpenRouter sends an empty `reasoning_details` array in the finish chunk;
-    // it must not wipe out details accumulated from earlier chunks.
-    #[test]
-    fn reasoning_details_accumulator_ignores_null_and_empty_array_chunks() {
-        let mut accumulator = ReasoningDetailsAccumulator::default();
-        assert!(
-            accumulator
-                .push(json!([{"index": 0, "type": "reasoning.text", "text": "thinking"}]))
-                .is_some()
-        );
-        assert!(
-            accumulator
-                .push(
-                    json!([{"index": 0, "type": "reasoning.encrypted", "data": "encrypted-blob"}])
-                )
-                .is_some()
-        );
-
-        assert_eq!(accumulator.push(json!([])), None);
-        assert_eq!(accumulator.push(serde_json::Value::Null), None);
-
-        let details = accumulator
-            .push(json!([{"index": 0, "text": " more"}]))
-            .expect("accumulated reasoning details");
-        assert_eq!(details[0]["text"], "thinking more");
-        assert_eq!(details[0]["data"], "encrypted-blob");
-    }
-
-    #[test]
-    fn stream_maps_preserves_tool_id_and_name_across_empty_deltas() {
-        // DashScope sends id="" and name="" in subsequent tool_calls delta
-        // chunks after the first chunk. OpenAiEventMapper must not overwrite
-        // the accumulated id and name with these empty strings.
-
-        let events = vec![
-            // First chunk: id and name are present
-            ResponseStreamEvent {
-                choices: vec![ChoiceDelta {
-                    index: 0,
-                    delta: Some(ResponseMessageDelta {
-                        content: None,
-                        reasoning: None,
-                        tool_calls: Some(vec![ToolCallChunk {
-                            index: 0,
-                            id: Some("call_dashscope_test".into()),
-                            function: Some(FunctionChunk {
-                                name: Some("list_directory".into()),
-                                arguments: Some("".into()),
-                                thought_signature: None,
-                            }),
-                        }]),
-                        reasoning_content: None,
-                        reasoning_details: None,
-                    }),
-                    finish_reason: None,
-                }],
-                usage: None,
-            },
-            // Subsequent chunks: DashScope sends id="" and name=""
-            ResponseStreamEvent {
-                choices: vec![ChoiceDelta {
-                    index: 0,
-                    delta: Some(ResponseMessageDelta {
-                        content: None,
-                        reasoning: None,
-                        tool_calls: Some(vec![ToolCallChunk {
-                            index: 0,
-                            id: Some("".into()),
-                            function: Some(FunctionChunk {
-                                name: Some("".into()),
-                                arguments: Some("{\"path\": \"".into()),
-                                thought_signature: None,
-                            }),
-                        }]),
-                        reasoning_content: None,
-                        reasoning_details: None,
-                    }),
-                    finish_reason: None,
-                }],
-                usage: None,
-            },
-            ResponseStreamEvent {
-                choices: vec![ChoiceDelta {
-                    index: 0,
-                    delta: Some(ResponseMessageDelta {
-                        content: None,
-                        reasoning: None,
-                        tool_calls: Some(vec![ToolCallChunk {
-                            index: 0,
-                            id: Some("".into()),
-                            function: Some(FunctionChunk {
-                                name: Some("".into()),
-                                arguments: Some("blog-scraper\"}".into()),
-                                thought_signature: None,
-                            }),
-                        }]),
-                        reasoning_content: None,
-                        reasoning_details: None,
-                    }),
-                    finish_reason: None,
-                }],
-                usage: None,
-            },
-            // Final chunk: finish_reason = "tool_calls"
-            ResponseStreamEvent {
-                choices: vec![ChoiceDelta {
-                    index: 0,
-                    delta: None,
-                    finish_reason: Some("tool_calls".into()),
-                }],
-                usage: None,
-            },
-        ];
-
-        let mapped = map_completion_events(events);
-
-        // Events emitted:
-        //   1. Partial ToolUse from chunk 1 (fix_json("") → "{}", parseable)
-        //   2. Partial ToolUse from chunk 3 (arguments fully assembled)
-        //   3. Complete ToolUse from finish_reason="tool_calls" drain
-        //   4. Stop(ToolUse)
-        assert_eq!(mapped.len(), 4);
-
-        // Verify the complete ToolUse event (from finish_reason drain)
-        // has the correct id, name, and accumulated arguments.
-        let complete_tool_use = mapped.iter().find_map(|event| {
-            if let LanguageModelCompletionEvent::ToolUse(tool_use) = event {
-                if tool_use.is_input_complete {
-                    return Some(tool_use);
-                }
-            }
-            None
-        });
-        assert!(
-            complete_tool_use.is_some(),
-            "expected a completed ToolUse event"
-        );
-        let tool_use = complete_tool_use.unwrap();
-        assert_eq!(
-            tool_use.id.to_string(),
-            "call_dashscope_test",
-            "id must survive empty-string overwrites"
-        );
-        assert_eq!(
-            tool_use.name.as_ref(),
-            "list_directory",
-            "name must survive empty-string overwrites"
-        );
-        assert_eq!(
-            tool_use.raw_input, "{\"path\": \"blog-scraper\"}",
-            "arguments should accumulate across chunks"
-        );
-
-        // Verify the Stop event
-        assert!(mapped.iter().any(|event| {
-            matches!(
-                event,
-                LanguageModelCompletionEvent::Stop(StopReason::ToolUse)
-            )
-        }));
     }
 
     #[test]
